Initial config for SKR1.3 with TMC2130 on Ender-3
This is a config with the BL-Touch plugged in the Z-min endstop for
homing with the Bl-Touch. This will be a problem if we want to get
the diagnostic from the Z axis (e.g. for crash detection) but this
can probably be solved by software (for latter).

This config only tested homing and bed leveling. I also tested
the SD-Card with Octoprint. I will do a test print tomorrow.