BMG clone adaptation
In addition to changing the direction of the motor and the esteps per
mm, this change set the hybrid threshold 3 times higher.

Since the bondtech needs 16x the steps of the old extruder to extrude
the same lenght of filament, it should be fine to stay in stealthchop
theoritically for 16x time the speed, so 3 time should be plenty and
that's 90 mm/s which should be way higher than any extrusion during
print.