Add filament sensor.
This is for a custom filament sensor. For some strange reason, even
though the filament sensor is normally open, it ends up being normally
closed (high) so high when the filament is not inserted ¯\_(ツ)_/¯.

I also changed the pin of the filament sensor to be Y+ instead of X+.
Indeed X+ is connected to the stallguard of the extruder which I want
to use to detect filament jam.