Update the Bl-Touch position
The previous extruder mount was terrible, updated it to https://www.thingiverse.com/thing:3856703

On the previous one, the fan shroud was hard to set the height correctly
and the fixation eventually broke, making the fan shroud fall down
during a print.

The new one is way nicer fit, the only drawback I see is that the
filament path is bigger.